backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
google-generativeai==0.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
